Refactor

+++ codekit-core/src/commons.rs
@@ -12,17 +12,17 @@ pub(crate) trait Barcode {
 #[repr(C)]
 #[derive(Debug, Clone, Copy)]
 pub struct CodeOptions {
-    pub quiet_space: f32,
-    pub code_height: f32,
-    pub border_width: f32,
+    pub quiet_space: u16,
+    pub code_height: u16,
+    pub border_width: u16,
 }
 
 impl Default for CodeOptions {
     fn default() -> Self {
         CodeOptions {
-            quiet_space: 7.0,
-            code_height: 50.0,
-            border_width: 0.0,
+            quiet_space: 7,
+            code_height: 50,
+            border_width: 0,
         }
     }
 }
@@ -39,15 +39,15 @@ impl Code {
         Code { options, bars }
     }
     /// Retrieve the border witdh of the code
-    pub fn border_width(&self) -> f32 {
+    pub fn border_width(&self) -> u16 {
         self.options.border_width
     }
 
-    pub fn height(&self) -> f32 {
+    pub fn height(&self) -> u16 {
         self.options.code_height
     }
 
-    pub fn quiet_space(&self) -> f32 {
+    pub fn quiet_space(&self) -> u16 {
         self.options.quiet_space
     }
